restructure the files and directories
