+++ service/web-server/src/execution/routes.ts
@@ -4,13 +4,14 @@
 
 import express from 'express';
 import { Client as IEsClient } from '@elastic/elasticsearch';
-import { ExecutionController, IExecutionController } from '../controllers/execution';
+import { ExecutionController, IExecutionController } from './controller';
+import { isPRMerged } from './middlewares';
 
 export const getExecutionRoutes = (esClient: IEsClient) => {
     const controller: IExecutionController = new ExecutionController(esClient);
     const router = express.Router();
 
-    router.post('/pull_request', controller.isPRMerged, controller.create);
+    router.post('/pull_request', isPRMerged, controller.create);
 
     router.post('/new', controller.create);
