+++ service/web-server/src/execution/vmClient/scripts/downloadLogs.sh
